Separate TypeScript utilities from React components
- Convert purchaseUtils.tsx to purchaseUtils.ts for pure TypeScript utilities
- Move StatusColorTooltip component to src/components/common/StatusColorTooltip.tsx
- Update import statements to reflect new file locations
- Improve code organization by separating utilities from UI components
- Maintain all existing functionality while following better project structure
